Simplified OS-specific fact setup, updated facts and added a README.md
Merged all tasks into tasks/main.yml for better maintenance.

Added fact 'distribution_id' which is a list that identifies a
distribution, e.g. [ 'Debian', '10' ] or [ 'Ubuntu', '20.04' ].

Changed fact 'distribution_codename' from a minus-separated string
to a list because this allows to join the list with a user-provided
separator, e.g. [ 'Debian', 'Buster' ] or [ 'Ubuntu', 'Focal' ].